Check the string palindrome or not
Problem statement
Write a function that takes a string as input and determines if it is a palindrome or not, considering only alphanumeric characters.

Note:
You are not required to print anything. It has already been taken care of. 
Detailed explanation ( Input/output format, Notes, Images )
Input Format:
The first and only line of input contains a string without any leading and trailing spaces. All the characters in the string would be in lower case.
Output Format:
The only line of output prints either 'Palindrome' or 'Not a Palindrome'.
Constraints:
0 <= N <= 10^6
Where N is the length of the input string.

Time Limit: 1 second
Sample Input 1 :
abcdcba
Sample Output 1 :
Palindrome 
Sample Input 2:
coding
Sample Output 2:
Not a Palindrome